rewrite: migrate `move_commits` function from `rebase` command
